+++ RedditClone.Application/ReplyVotes/Commands/DeleteReplyVote/DeleteReplyVoteCommandHandler.cs
@@ -40,6 +40,45 @@ public class DeleteReplyVoteCommandHandler(
             return error;
         }
 
+        if (!_commentRepository.UserExists(command.UserId))
+        {
+            Error error = Errors.User.UserNotFound;
+
+            Log.Error(
+                "{@Code}, {@Descriptor}",
+                error.Code,
+                error.Description);
+
+            return error;
+        }
+        var reply = comment.Replies.FirstOrDefault(r => r.Id == command.ReplyId);
+
+        if (reply is null)
+        {
+            Error error = Errors.CommentReplies.ReplyNotFound;
+
+            Log.Error(
+                "{@Code}, {@Descriptor}",
+                error.Code,
+                error.Description);
+
+            return error;
+        }
+
+        var vote = reply.Votes.FirstOrDefault(v => v.Id == command.VoteId);
+
+        if (vote is null)
+        {
+            Error error = Errors.ReplyVotes.VoteNotFound;
+
+            Log.Error(
+                "{@Code}, {@Descriptor}",
+                error.Code,
+                error.Description);
+
+            return error;
+        }
+
         _commentRepository.DeleteReplyVoteById(command.CommentId, command.ReplyId, command.VoteId, command.UserId);
 
         DeleteReplyVoteResult result = new("Vote deleted successfully.");
